Refactoring, moving missiles into their own package. Removing some redundant code.
