+++ server/src/main/kotlin/net/horizonsend/ion/server/features/starship/subsystem/weapon/projectile/missiles/LightMissileProjectile.kt
@@ -1,4 +1,4 @@
-package net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile
+package net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile.missiles
 
 import net.horizonsend.ion.common.utils.miscellaneous.randomDouble
 import net.horizonsend.ion.server.configuration.starship.StarshipTrackingProjectileBalancing
@@ -6,6 +6,8 @@ import net.horizonsend.ion.server.features.client.display.modular.ItemDisplayCon
 import net.horizonsend.ion.server.features.client.display.teleportDuration
 import net.horizonsend.ion.server.features.custom.items.util.ItemFactory
 import net.horizonsend.ion.server.features.starship.damager.Damager
+import net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile.SimpleProjectile
+import net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile.TrackingLaserProjectile
 import net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile.source.ProjectileSource
 import net.horizonsend.ion.server.miscellaneous.utils.coordinates.circlePoints
 import net.horizonsend.ion.server.miscellaneous.utils.coordinates.lerp
@@ -14,74 +16,37 @@ import org.bukkit.Color
 import org.bukkit.Location
 import org.bukkit.Particle
 import org.bukkit.block.BlockFace
-import org.bukkit.damage.DamageType
+import org.bukkit.entity.LivingEntity
 import org.bukkit.util.Vector
 
-class RapidHeavyMissileProjectile<B : StarshipTrackingProjectileBalancing>(
-    source: ProjectileSource,
-    name: Component,
-    loc: Location,
-    val dir: Vector,
-    val initialDir: Vector,
-    override val balancing: B,
-    shooter: Damager,
-    var face: BlockFace, //Up = true, down = false
-    originalTarget: Vector,
-    baseAimDistance: Int
-) : TrackingLaserProjectile<B>(source, name, loc, dir, shooter, originalTarget, baseAimDistance, DamageType.GENERIC) {
-	var flightPath1Completed = false
-	var flightPath2Completed = false
-	var age = 0
-	val delayMillis: Int = 150
-
-	val item = ItemFactory.Preset.unStackableCustomItem("projectile/activated_heavy_missile").construct()
+class LightMissileProjectile<B : StarshipTrackingProjectileBalancing>(
+	source: ProjectileSource,
+	name: Component,
+	loc: Location,
+	dir: Vector,
+	initialDir: Vector,
+	override val balancing: B,
+	shooter: Damager,
+	face: BlockFace, //Up = true, down = false
+	originalTarget: Vector,
+	baseAimDistance: Int
+) : PlayerGuidedLaserProjectile<B>(source, name, loc, dir, initialDir, balancing, shooter, face, originalTarget, baseAimDistance) {
+	override val item by lazy { ItemFactory.unStackableCustomItem("projectile/activated_light_missile").construct() }
 	override val color: Color = Color.ORANGE
+	override val verticalLaunchDistance = 10.0
 
-	init {
-		track = false
-	}
-
-	private val container = ItemDisplayContainer(
-        source.getWorld(),
-        2.0F,
-        loc.toVector(),
-        dir,
-        item,
-        interpolation = 2
-    ).apply {
-		getEntity().transformationInterpolationDuration = 2
-		getEntity().teleportDuration = 2
-	}
-
-	override fun tick() {
-
-		if (!flightPath1Completed) {
-			// Initial launch - fly straight out of the multiblock
-			direction =
-
-				initialDir.clone().multiply(1) // make the projectile launch parallel from the launcher, and slower
-
-			if (distance > 30) {
-				distance = 0.0
-				flightPath1Completed = true
-				track = true
-			}
-
-			super.tick()
-			return
-
-		} else if (!flightPath2Completed) {
-
-			// Transitioning to free flight - linearly interpolate between initial launch and free flight
-			direction = direction.clone().lerp(dir, 0.2)
-
-			super.tick()
-			return
+	override val container by lazy {
+		ItemDisplayContainer(
+			source.getWorld(),
+			2.0F,
+			loc.toVector(),
+			dir,
+			ItemFactory.unStackableCustomItem("projectile/activated_light_missile").construct(),
+			interpolation = 2
+		).apply {
+			getEntity().transformationInterpolationDuration = 2
+			getEntity().teleportDuration = 2
 		}
-
-		// Free flight
-		direction = dir
-
 	}
 
 	override fun moveVisually(oldLocation: Location, newLocation: Location, travel: Double) {
@@ -121,22 +86,6 @@ class RapidHeavyMissileProjectile<B : StarshipTrackingProjectileBalancing>(
 				null,
 				true
 			)
-		}
-	}
-
-
-	override var speed = balancing.speed
-
-	override fun onDespawn() {
-		container.remove()
-	}
-
-	private fun lerpAmount(distance: Double): Double {
-		val distanceRatio = distance / 67
-		return when {
-			distanceRatio <= 0.0 -> 1.0 // projectile is basically inside the target
-			distanceRatio >= 1.0 -> 0.0 // projectile is further than the proximity range
-			else -> /*1.00262 * 0.0511657.pow(distance / proximityRange)*/ 1 - distanceRatio
 		}
 	}
 
@@ -193,6 +142,5 @@ class RapidHeavyMissileProjectile<B : StarshipTrackingProjectileBalancing>(
 				true
 			)
 		}
-
 	}
 }
