+++ server/src/main/kotlin/net/horizonsend/ion/server/features/starship/subsystem/weapon/projectile/missiles/RapidHeavyMissileProjectile.kt
@@ -1,4 +1,4 @@
-package net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile
+package net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile.missiles
 
 import net.horizonsend.ion.common.utils.miscellaneous.randomDouble
 import net.horizonsend.ion.server.configuration.starship.StarshipTrackingProjectileBalancing
@@ -6,81 +6,44 @@ import net.horizonsend.ion.server.features.client.display.modular.ItemDisplayCon
 import net.horizonsend.ion.server.features.client.display.teleportDuration
 import net.horizonsend.ion.server.features.custom.items.util.ItemFactory
 import net.horizonsend.ion.server.features.starship.damager.Damager
+import net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile.TrackingLaserProjectile
 import net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile.source.ProjectileSource
 import net.horizonsend.ion.server.miscellaneous.utils.coordinates.circlePoints
 import net.horizonsend.ion.server.miscellaneous.utils.coordinates.lerp
 import net.kyori.adventure.text.Component
-import org.bukkit.damage.DamageType
 import org.bukkit.Color
 import org.bukkit.Location
 import org.bukkit.Particle
 import org.bukkit.block.BlockFace
+import org.bukkit.damage.DamageType
 import org.bukkit.util.Vector
 
-class LightMissileProjectile<B : StarshipTrackingProjectileBalancing>(
-	source: ProjectileSource,
-	name: Component,
-	loc: Location,
-	val dir: Vector,
+class RapidHeavyMissileProjectile<B : StarshipTrackingProjectileBalancing>(
+    source: ProjectileSource,
+    name: Component,
+    loc: Location,
+	dir: Vector,
 	initialDir: Vector,
-	override val balancing: B,
-	shooter: Damager,
-	var face: BlockFace, //Up = true, down = false
-	originalTarget: Vector,
-	baseAimDistance: Int
-) : PlayerGuidedLaserProjectile<B>(source, name, loc, dir, initialDir, balancing, shooter, originalTarget, baseAimDistance) {
-	var flightPath1Completed = false
-	var flightPath2Completed = false
-	var age = 0
+    override val balancing: B,
+    shooter: Damager,
+	face: BlockFace, //Up = true, down = false
+    originalTarget: Vector,
+    baseAimDistance: Int
+) : PlayerGuidedLaserProjectile<B>(source, name, loc, dir, initialDir, balancing, shooter, face, originalTarget, baseAimDistance) {
 
-	val item = ItemFactory.unStackableCustomItem("projectile/activated_light_missile").construct()
+	override val item = ItemFactory.Preset.unStackableCustomItem("projectile/activated_heavy_missile").construct()
 	override val color: Color = Color.ORANGE
 
-	init {
-		track = false
-	}
-
-	private val container = ItemDisplayContainer(
-		source.getWorld(),
-		2.0F,
-		loc.toVector(),
-		dir,
-		item,
-		interpolation = 2
-	).apply {
+	override val container = ItemDisplayContainer(
+        source.getWorld(),
+        2.0F,
+        loc.toVector(),
+        dir,
+        item,
+        interpolation = 2
+    ).apply {
 		getEntity().transformationInterpolationDuration = 2
 		getEntity().teleportDuration = 2
-	}
-
-	override fun tick() {
-
-		if (!flightPath1Completed) {
-			// Initial launch - fly straight out of the multiblock
-			direction =
-
-				initialDir.clone().multiply(1) // make the projectile launch parallel from the launcher, and slower
-
-			if (distance > 10) {
-				distance = 0.0
-				flightPath1Completed = true
-				track = true
-			}
-
-			super.tick()
-			return
-
-		} else if (!flightPath2Completed) {
-
-			// Transitioning to free flight - linearly interpolate between initial launch and free flight
-			direction = direction.clone().lerp(dir, 0.2)
-
-			super.tick()
-			return
-		}
-
-		// Free flight
-		direction = dir
-
 	}
 
 	override fun moveVisually(oldLocation: Location, newLocation: Location, travel: Double) {
@@ -123,18 +86,8 @@ class LightMissileProjectile<B : StarshipTrackingProjectileBalancing>(
 		}
 	}
 
-	override fun onDespawn() {
-		container.remove()
-	}
 
-	private fun lerpAmount(distance: Double): Double {
-		val distanceRatio = distance / 67
-		return when {
-			distanceRatio <= 0.0 -> 1.0 // projectile is basically inside the target
-			distanceRatio >= 1.0 -> 0.0 // projectile is further than the proximity range
-			else -> /*1.00262 * 0.0511657.pow(distance / proximityRange)*/ 1 - distanceRatio
-		}
-	}
+	override var speed = balancing.speed
 
 	override fun onImpact() {
 		for (loc in location.circlePoints(2.0, 30, direction)) {
@@ -189,5 +142,6 @@ class LightMissileProjectile<B : StarshipTrackingProjectileBalancing>(
 				true
 			)
 		}
+
 	}
 }
