+++ server/src/main/kotlin/net/horizonsend/ion/server/features/starship/subsystem/weapon/projectile/missiles/TrackingMissileProjectile.kt
@@ -1,28 +1,25 @@
-package net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile
+package net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile.missiles
 
 import net.horizonsend.ion.common.utils.miscellaneous.randomDouble
-import net.horizonsend.ion.common.utils.miscellaneous.randomInt
 import net.horizonsend.ion.server.configuration.starship.StarshipTrackingProjectileBalancing
 import net.horizonsend.ion.server.features.client.display.modular.ItemDisplayContainer
 import net.horizonsend.ion.server.features.client.display.teleportDuration
 import net.horizonsend.ion.server.features.custom.items.util.ItemFactory
 import net.horizonsend.ion.server.features.starship.damager.Damager
+import net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile.TrackingLaserProjectile
 import net.horizonsend.ion.server.features.starship.subsystem.weapon.projectile.source.ProjectileSource
-import net.horizonsend.ion.server.miscellaneous.utils.Tasks
 import net.horizonsend.ion.server.miscellaneous.utils.coordinates.circlePoints
 import net.horizonsend.ion.server.miscellaneous.utils.coordinates.lerp
-import net.horizonsend.ion.server.miscellaneous.utils.coordinates.spherePoints
 import net.kyori.adventure.text.Component
 import org.bukkit.damage.DamageType
 import org.bukkit.Color
 import org.bukkit.Location
 import org.bukkit.Particle
-import org.bukkit.block.Block
 import org.bukkit.block.BlockFace
-import org.bukkit.entity.Entity
+import org.bukkit.inventory.ItemStack
 import org.bukkit.util.Vector
 
-class ThermonuclearMissileProjectile<B : StarshipTrackingProjectileBalancing>(
+abstract class TrackingMissileProjectile<B : StarshipTrackingProjectileBalancing>(
 	source: ProjectileSource,
 	name: Component,
 	loc: Location,
@@ -33,34 +30,29 @@ class ThermonuclearMissileProjectile<B : StarshipTrackingProjectileBalancing>(
 	var face: BlockFace, //Up = true, down = false
 	originalTarget: Vector,
 	baseAimDistance: Int
-) : TrackingLaserProjectile<B>(source, name, loc, dir, shooter, originalTarget, baseAimDistance, DamageType.GENERIC) {
+) : TrackingLaserProjectile<B>(source, name, loc, dir, shooter, originalTarget, baseAimDistance,  DamageType.GENERIC) {
 	var flightPath1Completed = false
 	var flightPath2Completed = false
-	var age = 0
-	private val explosionParticleData = Particle.DustTransition(
-		Color.AQUA,
-		Color.WHITE,
-		balancing.particleThickness.toFloat()
-	)
+	open val verticalLaunchDistance = 30.0
+	override var track = false
 
-
-	val item = ItemFactory.unStackableCustomItem("projectile/activated_thermonuclear_missile").construct()
-	override val color: Color = Color.ORANGE
-
-	init {
-		track = false
+	open val item by lazy { ItemFactory.unStackableCustomItem("projectile/activated_arsenal_missile").construct()}
+	open val container by lazy {
+		ItemDisplayContainer(
+			source.getWorld(),
+			4.0F,
+			loc.toVector(),
+			dir.clone().multiply(-1),
+			ItemFactory.unStackableCustomItem("projectile/activated_arsenal_missile").construct(),
+			interpolation = 2
+		).apply {
+			getEntity().transformationInterpolationDuration = 2
+			getEntity().teleportDuration = 2
+		}
 	}
 
-	private val container = ItemDisplayContainer(
-		source.getWorld(),
-		6.0F,
-		loc.toVector(),
-		dir,
-		item,
-		interpolation = 2
-	).apply {
-		getEntity().transformationInterpolationDuration = 2
-		getEntity().teleportDuration = 2
+	init {
+	    track = false
 	}
 
 	override fun tick() {
@@ -71,7 +63,7 @@ class ThermonuclearMissileProjectile<B : StarshipTrackingProjectileBalancing>(
 
 				initialDir.clone().multiply(1) // make the projectile launch parallel from the launcher, and slower
 
-			if (distance > 30) {
+			if (distance > verticalLaunchDistance) {
 				distance = 0.0
 				flightPath1Completed = true
 				track = true
@@ -148,81 +140,58 @@ class ThermonuclearMissileProjectile<B : StarshipTrackingProjectileBalancing>(
 	}
 
 	override fun onImpact() {
-		val task = Tasks.syncRepeatTask(0L, 2L) {
-			for (point in location.circlePoints(3.0, 10, direction)) {
-				point.world.spawnParticle(
-					Particle.DUST_COLOR_TRANSITION,
-					point.x,
-					point.y,
-					point.z,
-					1,
-					0.0,
-					0.0,
-					0.0,
-					2.0,
-					explosionParticleData,
-					true
-				)
-			}
-			val endLocation = location.toCenterLocation()
-
-			for (startPoint in location.circlePoints(5.0, 20, direction)) {
-				location.world.spawnParticle(
-					Particle.TRAIL,
-					startPoint,
-					1,
-					0.5,
-					0.5,
-					0.5,
-					0.0,
-					Particle.Trail(endLocation, Color.AQUA, randomInt(19, 21)),
-					true
-				)
-			}
+		for (loc in location.circlePoints(2.0, 30, direction)) {
+			val radialVector = loc.toVector().subtract(location.toVector()).normalize()
+			loc.world.spawnParticle(
+				Particle.CLOUD,
+				location,
+				0,
+				radialVector.x,
+				radialVector.y,
+				radialVector.z,
+				1.0,
+				null,
+				true
+			)
 		}
 
-		Tasks.syncDelay(40L) {
-			task.cancel()
+		(0 until 20).forEach { _ ->
+			val angle = Math.PI / 12
+			val opposite = direction.clone().multiply(-1)
+				.rotateAroundX(randomDouble(-angle, angle))
+				.rotateAroundY(randomDouble(-angle, angle))
+				.rotateAroundZ(randomDouble(-angle, angle))
+			location.world.spawnParticle(
+				Particle.FLAME,
+				location,
+				0,
+				opposite.x,
+				opposite.y,
+				opposite.z,
+				1.0,
+				null,
+				true
+			)
 		}
 
-		Tasks.syncDelay(40L) {
-			(0 until 20).forEach { _ ->
-				val angle = Math.PI / 12
-				val opposite = direction.clone().multiply(-1)
-					.rotateAroundX(randomDouble(-angle, angle))
-					.rotateAroundY(randomDouble(-angle, angle))
-					.rotateAroundZ(randomDouble(-angle, angle))
-				location.world.spawnParticle(
-					Particle.SOUL_FIRE_FLAME,
-					location,
-					0,
-					opposite.x,
-					opposite.y,
-					opposite.z,
-					1.0,
-					null,
-					true
-				)
-			}
-
-			(0 until 40).forEach { _ ->
-				val angle = Math.PI / 6
-				val opposite = direction.clone().multiply(-1)
-					.rotateAroundX(randomDouble(-angle, angle))
-					.rotateAroundY(randomDouble(-angle, angle))
-					.rotateAroundZ(randomDouble(-angle, angle))
-				location.world.spawnParticle(
-					Particle.SOUL_FIRE_FLAME,
-					location,
-					0,
-					opposite.x,
-					opposite.y,
-					opposite.z,
-					2.0,
-					null,
-					true
-				)
-			}
+		(0 until 40).forEach { _ ->
+			val angle = Math.PI / 6
+			val opposite = direction.clone().multiply(-1)
+				.rotateAroundX(randomDouble(-angle, angle))
+				.rotateAroundY(randomDouble(-angle, angle))
+				.rotateAroundZ(randomDouble(-angle, angle))
+			location.world.spawnParticle(
+				Particle.LARGE_SMOKE,
+				location,
+				0,
+				opposite.x,
+				opposite.y,
+				opposite.z,
+				2.0,
+				null,
+				true
+			)
 		}
+
 	}
 }
